fix(ui): make bar delta arithmetic GCC-clean and overflow-safe

+++ firmware/leshy1/src/ui/LayeredBarDelta.h
@@ -22,22 +22,27 @@ void paintLayeredBarDelta(const BarLayers& old, const BarLayers& next,
         return;
     if (old.current == next.current && old.average == next.average &&
         old.color == next.color) return;
-    const std::int16_t inset = (width - currentWidth) / 2;
+    // The validated widths keep these differences within int16_t.
+    const std::int16_t inset =
+        static_cast<std::int16_t>((width - currentWidth) / 2);
     const std::int16_t edges[] = {0, inset,
         static_cast<std::int16_t>(inset + currentWidth), width};
     for (unsigned strip = 0; strip < 3; ++strip) {
-        const std::int16_t span = edges[strip + 1U] - edges[strip];
+        const std::int16_t span =
+            static_cast<std::int16_t>(edges[strip + 1U] - edges[strip]);
         if (span == 0) continue;
         std::int16_t run = -1;
         std::uint16_t runColor = 0;
-        for (std::int16_t y = 0; y <= height; ++y) {
+        // One terminal sentinel row; a 16-bit iterator would wrap at INT16_MAX.
+        for (std::int32_t y = 0; y <= height; ++y) {
             bool changed = false;
             std::uint16_t color = 0;
             if (y < height) {
                 const auto pixel = [&](const BarLayers& bar) {
                     if (strip == 1U && y >= height - bar.current) return bar.color;
                     if (y >= height - bar.average) return averageColor;
-                    return static_cast<std::uint16_t>(background(y));
+                    return static_cast<std::uint16_t>(
+                        background(static_cast<std::int16_t>(y)));
                 };
                 color = pixel(next);
                 changed = color != pixel(old);
@@ -47,7 +52,9 @@ void paintLayeredBarDelta(const BarLayers& old, const BarLayers& next,
                       static_cast<std::int16_t>(y - run), runColor);
                 run = -1;
             }
-            if (changed && run < 0) { run = y; runColor = color; }
+            if (changed && run < 0) {
+                run = static_cast<std::int16_t>(y); runColor = color;
+            }
         }
     }
 }
